fix(*): handle possible exception accessing writeStream
If the client misbehaves an undefined upload can end up in the server,
leading it to a crash because no reference exists to it.  As this
would throw an exception inside a callback the exception isn't handled
properly ending up with a crash.  Use a try/catch to avoid that
unhandled exception.

Add package-lock.json.